feat(kernel): spilt basic_info as a file
