+++ src/kernel/arch/x86_64/arch_main.cpp
@@ -16,10 +16,10 @@
 
 #include <elf.h>
 
+#include "basic_info.hpp"
 #include "cpu.hpp"
 #include "cstdio"
 #include "cstring"
-#include "kernel.h"
 #include "kernel_elf.hpp"
 #include "kernel_log.hpp"
